+++ include/direwolf/renderengine.h
@@ -11,6 +11,10 @@ public:
     RenderEngine(const PlatformData& platformData, const InitData& initData);
 
 	bool CreateVertexBuffer(const GfxObject& object, uint32_t count) const;
+	void* MapVertexBuffer(const GfxObject& object);
+	void UnmapVertexBuffer(const GfxObject& object);
+	void DestroyVertexBuffer(const GfxObject& object);
+
 	bool CreatePipelineState(const GfxObject& object, const PipelineState& pipelineState) const;
 	void Render(const std::vector<RenderCommand>& commandBuffer) const;
