+++ src/renderengine.cpp
@@ -23,6 +23,14 @@ bool RenderEngine::CreateVertexBuffer(const GfxObject& object, uint32_t count) c
 	return m_renderer->CreateVertexBuffer(object, count);
 }
 
+void* RenderEngine::MapVertexBuffer(const GfxObject& object) {
+	return m_renderer->MapVertexBuffer(object);
+}
+
+void RenderEngine::UnmapVertexBuffer(const GfxObject& object) {
+    m_renderer->UnmapVertexBuffer(object);
+}
+
 bool RenderEngine::CreatePipelineState(const GfxObject& object, const PipelineState& pipelineState) const {
 	return m_renderer->CreatePipelineState(object, pipelineState);
 }
